+++ engine/flink/management/src/it/scala/pl/touk/nussknacker/engine/management/streaming/FlinkStreamingDeploymentManagerSlotsCountSpec.scala
@@ -1,16 +1,13 @@
 package pl.touk.nussknacker.engine.management.streaming
 
 import org.scalatest.{FunSuite, Matchers}
-import pl.touk.nussknacker.engine.api.deployment.{DeploymentData, GraphProcess, StateStatus}
+import pl.touk.nussknacker.engine.api.ProcessVersion
+import pl.touk.nussknacker.engine.api.deployment.{DeploymentData, GraphProcess}
 import pl.touk.nussknacker.engine.api.process.ProcessName
-import pl.touk.nussknacker.engine.api.{ProcessVersion, StreamMetaData}
 import pl.touk.nussknacker.engine.canonize.ProcessCanonizer
-import pl.touk.nussknacker.engine.graph.EspProcess
-import pl.touk.nussknacker.engine.management.FlinkStateStatus
+import pl.touk.nussknacker.engine.management.FlinkSlotsChecker.{NotEnoughSlotsException, SlotsBalance}
 import pl.touk.nussknacker.engine.marshall.ProcessMarshaller
 import pl.touk.nussknacker.engine.util.config.ScalaMajorVersionConfig
-
-import scala.concurrent.duration._
 
 class FlinkStreamingDeploymentManagerSlotsCountSpec extends FunSuite with Matchers with StreamingDockerTest {
 
@@ -18,29 +15,19 @@ class FlinkStreamingDeploymentManagerSlotsCountSpec extends FunSuite with Matche
 
   override val taskManagerSlotCount: Int = 1
 
-  // manual test because it takes a while to verify that
-  ignore("deploy scenario with too low task manager slots counts") {
+  test("deploy scenario with too low task manager slots counts") {
     val processId = "processTestingTMSlots"
     val version = ProcessVersion(1, ProcessName(processId), "user1", Some(13))
-    val process = SampleProcess.prepareProcess(processId, parallelism = Some(2))
+    val parallelism = 2
+    val process = SampleProcess.prepareProcess(processId, parallelism = Some(parallelism))
 
     try {
-      deployProcess(process, version)
-      continuouslyHaveStateStatus(process.id, FlinkStateStatus.DuringDeploy, 10, 1 second)
+      val marshaled = ProcessMarshaller.toJson(ProcessCanonizer.canonize(process)).spaces2
+      deploymentManager.deploy(version, DeploymentData.empty, GraphProcess(marshaled), None).failed.futureValue shouldEqual
+        NotEnoughSlotsException(taskManagerSlotCount, taskManagerSlotCount, SlotsBalance(0, parallelism))
     } finally {
       cancelProcess(processId)
     }
   }
 
-  private def continuouslyHaveStateStatus(processId: String, expectedStatus: StateStatus, attempts: Int, checkInterval: FiniteDuration): Unit = {
-    (0 until attempts).foreach { attempt =>
-      val jobStatus = deploymentManager.findJobStatus(ProcessName(processId)).futureValue
-      logger.debug(s"Checking if: $processId, have expected state status: $expectedStatus, current job status: $jobStatus, attempt: $attempt")
-
-      jobStatus.map(_.status.name) shouldBe Some(expectedStatus.name)
-      if (attempt != attempts - 1) {
-        Thread.sleep(checkInterval.toMillis)
-      }
-    }
-  }
 }