+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkRestManager.scala
@@ -2,13 +2,14 @@ package pl.touk.nussknacker.engine.management
 
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.flink.api.common.JobStatus
+import org.apache.flink.configuration.Configuration
 import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.namespaces.{FlinkUsageKey, NamingContext}
 import pl.touk.nussknacker.engine.api.process.ProcessName
-import pl.touk.nussknacker.engine.management.rest.HttpFlinkClient
-import pl.touk.nussknacker.engine.management.rest.flinkRestModel.JobOverview
+import pl.touk.nussknacker.engine.management.rest.{HttpFlinkClient, flinkRestModel}
+import pl.touk.nussknacker.engine.management.rest.flinkRestModel.{ClusterOverview, JobOverview}
 import sttp.client._
 
 import java.io.File
@@ -25,6 +26,8 @@ class FlinkRestManager(config: FlinkConfig, modelData: ModelData, mainClassName:
   protected lazy val jarFile: File = new FlinkModelJar().buildJobJar(modelData)
 
   private val client = new HttpFlinkClient(config)
+
+  private val slotsChecker = new FlinkSlotsChecker(client)
 
   /*
   It's ok to have many jobs with same name, however:
@@ -167,6 +170,10 @@ class FlinkRestManager(config: FlinkConfig, modelData: ModelData, mainClassName:
   override protected def runProgram(processName: ProcessName, mainClass: String, args: List[String], savepointPath: Option[String]): Future[Option[ExternalDeploymentId]] = {
     logger.debug(s"Starting to deploy scenario: $processName with savepoint $savepointPath")
     client.runProgram(jarFile, mainClass, args, savepointPath)
+  }
+
+  override protected def checkRequiredSlotsExceedAvailableSlots(processDeploymentData: ProcessDeploymentData, currentlyDeployedJobId: Option[ExternalDeploymentId]): Future[Unit] = {
+    slotsChecker.checkRequiredSlotsExceedAvailableSlots(processDeploymentData, currentlyDeployedJobId)
   }
 
   override def close(): Unit = Await.result(backend.close(), Duration(10, TimeUnit.SECONDS))