+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/rest/flinkRestModel.scala
@@ -48,13 +48,17 @@ object flinkRestModel {
 
   @JsonCodec(decodeOnly = true) case class JobConfig(jid: String, `execution-config`: ExecutionConfig)
 
-  @JsonCodec(decodeOnly = true) case class ExecutionConfig(`user-config`: Map[String, io.circe.Json])
+  @JsonCodec(decodeOnly = true) case class ExecutionConfig(`job-parallelism`: Int, `user-config`: Map[String, io.circe.Json])
 
   @JsonCodec(decodeOnly = true) case class JarsResponse(files: Option[List[JarFile]])
 
   @JsonCodec(decodeOnly = true) case class UploadJarResponse(filename: String)
 
   @JsonCodec(decodeOnly = true) case class JarFile(id: String, name: String)
+
+  @JsonCodec(decodeOnly = true) case class ClusterOverview(`slots-total`: Int, `slots-available`: Int)
+
+  @JsonCodec(decodeOnly = true) case class KeyValueEntry(key: String, value: String)
 
   @JsonCodec case class RunResponse(jobid: String)